refactor(capabilities): :recycle: migrate content/art to component-store pattern
Flatten art/components/ into per-component folders with store hooks.
- art/: pure page component, no store
- art-header/: pure presentational, no store
- art-gallery/: stateful, useArtGalleryStore (currentImage state, curried selectImage/clearImage effects)
- art-gallery/modal-with-image/: nested (single consumer = art-gallery), pure presentational
