+++ src/components/content/art/art-gallery/index.ts
@@ -1,0 +1,1 @@
+export { ArtGallery } from "./art-gallery";
