+++ src/components/content/art/art-gallery/modal-with-image/index.ts
@@ -1,0 +1,2 @@
+export { ModalWithImage } from "./modal-with-image";
+export type { ModalWithImageProps } from "./modal-with-image";
