+++ src/components/content/art/art-header/index.ts
@@ -1,0 +1,1 @@
+export { ArtHeader } from "./art-header";
