+++ src/components/content/art/art/index.ts
@@ -1,0 +1,1 @@
+export { Art } from "./art";
