--- src/components/content/art/components/art.tsx
@@ -1,17 +1,0 @@
-import { ContentPage } from "@/components/features/content/content-page";
-import { siteMetadata } from "@/types/configuration/site-metadata";
-import { tracking } from "@/types/configuration/tracking";
-import { ArtGallery } from "./art-gallery";
-import { ArtHeader } from "./art-header";
-
-export const Art = () => {
-  return (
-    <ContentPage
-      author={siteMetadata.author}
-      trackingCategory={tracking.category.art}
-    >
-      <ArtHeader />
-      <ArtGallery />
-    </ContentPage>
-  );
-};
